dom/tests/mochitest/geolocation/test_worseAccuracyDoesNotBlockCallback.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=494924
-->
<head>
  <title>Test for getCurrentPosition </title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="geolocation_common.js"></script>

<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=494924">Mozilla Bug 494924</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
<script class="testbody" type="text/javascript">
"use strict";

SimpleTest.waitForExplicitFinish();

resume_geolocationProvider(function() {
  force_prompt(true, test1);
});

// The magic numbers 100 and 42 below are the accuracy results returned from the
// mock geolocation service and defined in network_geolocation.sjs.
function successCallback2(position) {
  check_geolocation(position);
  is(position.coords.accuracy, 100, "worse accuracy");
  SimpleTest.finish();
}

function successCallback1(position) {
  check_geolocation(position);
  is(position.coords.accuracy, 42, "high accuracy");
  worse_geolocationProvider(function() {
    navigator.geolocation.getCurrentPosition(successCallback2, null, { maximumAge: 0 });
  });
}

function test1() {
  navigator.geolocation.getCurrentPosition(successCallback1, null, { maximumAge: 0 });
}
</script>
</pre>
</body>
</html>
